requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple
--extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple
--extra-index-url https://pypi.org/simple

aiohttp==3.9.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
appdirs==1.4.4
    # via wandb
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   wandb
cnocr[ort-cpu]==2.3.0.2
    # via
    #   -r requirements.in
    #   cnocr
cnstd==1.2.4.2
    # via
    #   -r requirements.in
    #   cnocr
coloredlogs==15.0.1
    # via
    #   onnxruntime
    #   optimum
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==2.17.0
    # via
    #   evaluate
    #   optimum
dill==0.3.8
    # via
    #   datasets
    #   evaluate
    #   multiprocess
docker-pycreds==0.4.0
    # via wandb
easyocr==1.7.1
    # via -r requirements.in
evaluate==0.4.1
    # via optimum
filelock==3.13.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==23.5.26
    # via onnxruntime
fonttools==4.49.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.10.0
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via wandb
huggingface-hub==0.20.3
    # via
    #   cnstd
    #   datasets
    #   evaluate
    #   optimum
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.6
    # via
    #   requests
    #   yarl
imageio==2.34.0
    # via scikit-image
importlib-resources==6.1.1
    # via matplotlib
jinja2==3.1.3
    # via torch
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
lightning-utilities==0.10.1
    # via
    #   pytorch-lightning
    #   torchmetrics
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via
    #   cnstd
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   datasets
    #   evaluate
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
ninja==1.11.1.1
    # via easyocr
numpy==1.26.4
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   contourpy
    #   datasets
    #   easyocr
    #   evaluate
    #   imageio
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   opencv-python-headless
    #   optimum
    #   pandas
    #   pyarrow
    #   pytorch-lightning
    #   scikit-image
    #   scipy
    #   seaborn
    #   shapely
    #   tifffile
    #   torchmetrics
    #   torchvision
    #   transformers
    #   ultralytics-thop
# onnx==1.15.0
    # via
    #   cnocr
    #   cnstd
    #   optimum
# onnxruntime==1.17.0
    # via
    #   cnocr
    #   optimum
opencv-python==4.9.0.80
    # via
    #   -r requirements.in
    #   cnstd
    #   ultralytics
opencv-python-headless==4.9.0.80
    # via easyocr
# optimum[onnxruntime]==1.16.2
    # via -r requirements.in
packaging==23.2
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   onnxruntime
    #   optimum
    #   pytorch-lightning
    #   scikit-image
    #   torchmetrics
    #   transformers
pandas==2.2.0
    # via
    #   cnstd
    #   datasets
    #   evaluate
    #   seaborn
    #   ultralytics
pillow==10.2.0
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   easyocr
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
    #   ultralytics
polygon3==3.0.9.1
    # via cnstd
protobuf==4.25.3
    # via
    #   onnx
    #   onnxruntime
    #   optimum
    #   transformers
    #   wandb
psutil==5.9.8
    # via
    #   ultralytics
    #   wandb
py-cpuinfo==9.0.0
    # via ultralytics
pyarrow==15.0.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pyclipper==1.3.0.post5
    # via
    #   cnstd
    #   easyocr
pymupdf==1.24.1
    # via -r requirements.in
pymupdfb==1.24.1
    # via pymupdf
pyparsing==3.1.1
    # via matplotlib
pyspellchecker==0.8.1
    # via -r requirements.in
python-bidi==0.4.2
    # via easyocr
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==2.2.0.post0
    # via
    #   cnocr
    #   cnstd
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   cnstd
    #   datasets
    #   easyocr
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
    #   ultralytics
    #   wandb
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   datasets
    #   evaluate
    #   fsspec
    #   huggingface-hub
    #   responses
    #   torchvision
    #   transformers
    #   ultralytics
    #   wandb
responses==0.18.0
    # via evaluate
safetensors==0.4.2
    # via transformers
scikit-image==0.22.0
    # via easyocr
scipy==1.12.0
    # via
    #   cnstd
    #   easyocr
    #   scikit-image
    #   ultralytics
seaborn==0.13.2
    # via
    #   cnstd
    #   ultralytics
sentencepiece==0.1.99
    # via transformers
sentry-sdk==1.40.4
    # via wandb
setproctitle==1.3.3
    # via wandb
shapely==2.0.2
    # via
    #   cnstd
    #   easyocr
six==1.16.0
    # via
    #   docker-pycreds
    #   python-bidi
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sympy==1.12
    # via
    #   onnxruntime
    #   optimum
    #   torch
tifffile==2024.2.12
    # via scikit-image
tokenizers==0.15.2
    # via transformers
torch==2.2.0
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   easyocr
    #   optimum
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchmetrics==1.3.1
    # via
    #   cnocr
    #   pytorch-lightning
torchvision==0.17.0
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   easyocr
    #   ultralytics
tqdm==4.66.2
    # via
    #   -r requirements.in
    #   cnocr
    #   cnstd
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
    #   ultralytics
transformers[sentencepiece]==4.37.2
    # via
    #   -r requirements.in
    #   optimum
typing-extensions==4.9.0
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
    #   wandb
tzdata==2024.1
    # via pandas
ultralytics==8.2.32
    # via cnstd
ultralytics-thop==0.2.8
    # via ultralytics
unidecode==1.3.8
    # via cnstd
urllib3==2.2.0
    # via
    #   requests
    #   responses
    #   sentry-sdk
wandb==0.16.3
    # via cnocr
xxhash==3.4.1
    # via
    #   datasets
    #   evaluate
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
